Added a few trivial functions for testing

+++ src/main/java/edu/jhu/hlt/optimize/RealScalarFunction.java
@@ -32,5 +32,5 @@ public interface RealScalarFunction extends Function {
 	/**
 	 * @return		The number of (free) variables
 	 */
-	double dim();
+	int dim();
 }
